Added wordlist manager genserver
