+++ test/test_helper.exs
@@ -1,1 +1,3 @@
 ExUnit.start()
+
+{:ok, _pid} = NotQwerty123.WordlistManager.start_link
